requirements.txt
flask
sentence-transformers
faiss-cpu
PyMuPDF
transformers
torch

